better rails useful

+++ rails/rails_useful.rb
@@ -14,16 +14,21 @@ end
 generic_patches = File.dirname(__FILE__) + '/rails_generic.rb'
 require generic_patches if File.exist?(generic_patches)
 
-if (ENV['RAILS_ENV'] == 'production' or ENV['RAILS_ENV'] == 'staging') and Socket.gethostname =~ /Rogers/ # ruby does it itself otherwise, I think.  There may be a rails way to do this.
+
+# takes dirs relative to rails' root
+
+def fire_up_file_modification_checker_thread reload_dirs = ['app/controllers', 'app/schools', 'app/models'], death_dirs =  ['app/helpers', 'config', 'app/controllers/shared_search_code_module.rb'] # actually these dirs are arrays of dirs, and can also include specific filenames
+
 watcher_thread = Thread.new{
+begin
 print 'STARTING WATCHER'
 latest_inserted = Time.now
-reload_dirs = ['app/controllers', 'app/schools', 'app/models']
-death_dirs = ['app/helpers', 'config']
 loop do
  for dir in death_dirs do
-	for file in (Dir.glob dir + '/*') + (Dir.glob dir + '/*/*')
+	for file in (Dir.glob dir + '/*') + (Dir.glob dir + '/*/*') + [dir]
+		next unless File.exist? file
 		time = File.ctime file
+		time  ||= -1
 		if time > latest_inserted
 			print 'got new' , file, "\n KILLING\n"
   			system("kill -9 #{Process.pid}") # we are done
@@ -34,6 +39,7 @@ loop do
  has_new = false
  for dir in reload_dirs
 	for file in (Dir.glob dir + '/*') + (Dir.glob dir + '/*/*')
+		next unless File.exist? file
   
 		time = File.ctime file
 		if time > latest_inserted
@@ -67,5 +73,15 @@ loop do
  end
  sleep 0.2
  end
+ rescue Exception => e
+   print "ARRRR\n\n\n\n", e, e.backtrace.join("\n")
+   raise
+ end
 }
 end
+
+
+if (ENV['RAILS_ENV'] == 'production' or ENV['RAILS_ENV'] == 'staging') and Socket.gethostname =~ /roger/ # ruby does it itself otherwise, I think.  There may be a rails way to do this.
+ fire_up_file_modification_checker_thread
+end
+
